iota_sdk_grpc_types::v1

Module transaction

Modules

object_change

Nested message and enum types in ObjectChange.

Structs

BalanceChange

The delta, or change, in balance of a particular coin type for an owner.

BalanceChangeFieldPathBuilder

BalanceChanges

A list of balance changes.

BalanceChangesFieldPathBuilder

ExecutedTransaction

A transaction that has been executed, along with its signatures, effects, events and objects.

ExecutedTransactionFieldPathBuilder

ExecutedTransactions

ExecutedTransactionsFieldPathBuilder

ObjectChange

A change to an object caused by executing a transaction.

ObjectChangeCreated

A new object was created.

ObjectChangeCreatedFieldPathBuilder

ObjectChangeDeleted

An object was deleted.

ObjectChangeDeletedFieldPathBuilder

ObjectChangeFieldPathBuilder

ObjectChangeMutated

An object was mutated.

ObjectChangeMutatedFieldPathBuilder

ObjectChangePublished

A package was published.

ObjectChangePublishedFieldPathBuilder

ObjectChangeUnwrapped

An object was unwrapped from inside another object.

ObjectChangeUnwrappedFieldPathBuilder

ObjectChangeWrapped

An object was wrapped inside another object.

ObjectChangeWrappedFieldPathBuilder

ObjectChanges

A list of object changes.

ObjectChangesFieldPathBuilder

Transaction

A transaction.

TransactionEffects

The effects of executing a transaction.

TransactionEffectsFieldPathBuilder

TransactionEvents

iota.grpc.v1.event.Events emitted during the successful execution of a transaction.

TransactionEventsFieldPathBuilder

TransactionFieldPathBuilder